processor_prime

+++ procc/assembller/file_commands.h
@@ -4,7 +4,7 @@
 #include "enumchik_s.h"
 #include "str_of_str.h"
 
-char** glue_of_functio(int& count_operation, int& lines_num);
+char** glue_of_functio(size_t& count_operation, int& lines_num);
 
 char* file_pointer_read();
