+++ procc/assembller/str_of_str.h
@@ -3,6 +3,6 @@
 
 #include"enumchik_s.h"
 
-char** v_str_ku(char *full_asm, int lines_num, size_t number_char_in_file, int& count_operation);
+char** v_str_ku(char *full_asm, int lines_num, size_t number_char_in_file, size_t& count_operation);
 
 #endif
